### 2020-07-01 #2418 @whedon
[PRE REVIEW]: chirps: API Client for the CHIRPS Precipitation Data in R
**Submitting author:** @kauedesousa (<a href="http://orcid.org/0000-0002-7571-7845">Kauê de Sousa</a>)
**Repository:** <a href="https://github.com/ropensci/chirps" target ="_blank">https://github.com/ropensci/chirps</a>
**Version:** v0.1.0
