### 2020-07-01 #2418 @kauedesousa
This package has been reviewed by rOpenSci: https://github.com/ropensci/software-review/issues/357, @melvidoni
