### 2020-07-01 #2419 @danielskatz
This is an expedited review, as this package has been reviewed by rOpenSci: https://github.com/ropensci/software-review/issues/357
