### 2020-07-01 #2419 @danielskatz
👋 @kauedesousa - In the .bib file, can you protect the cases (using {}s) in the proceedings title for Kehel et al. so that it is produced in Title Case rather than Sentence case?
