### 2020-07-01 #2419 @kauedesousa
Hi @danielskatz the title cases were protected and pushed here https://github.com/ropensci/chirps/commit/34747803b8c673be225f30338c10feb3f04bc211
