### 2020-07-01 #2419 @danielskatz
@kauedesousa - sorry, a few more than I see - can you fix
- Mozambique in Horn
- JSON and R in Ooms
Please regenerate the pdf after making changes by ```@whedon generate pdf```
