### 2020-07-01 #2419 @danielskatz
I don't yet see the paper - once it is there, I will close this issue and we will be done.
